+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `RURL.isUrl()` should no longer return true for objects containing just a `href` property
 * The `body` of a `Request` can now also be set with a `body` option property
 * Fix `RURL#param(name, value)` not parsing form paths
+* Add some basic timezone support to `Date#format(pattern, locale, timezone)`
 
 ## 0.7.4 (2020-10-22)